fix: Use correct type when connecting to Redis Cluster
Connecting to a redis cluster using Redis runs into a variety of issues.
For example, it will not follow MOVED errors. In order to get desired
behavior, use the Redis.Cluster type. This requires a reworking of how
we configure the server to connect to Redis because of differences in
constructing a Redis instance vs a Redis.Cluster instance.